demo: Fix sporadic bad test results
Sometimes when running the demo test, incorrect results are occasionally
seen when querying the movie count after increasing the rating of
"Battlefield Earth".  It's possible that this is some kind of
replication lag.  Sleeping for a few seconds between CR presses during
the demo seems to give the setup time to adjust, and reduces occurrence
of bad results.

Change-Id: I6a6a69641c88d00eb10ae6df9fbe74af1228ad47
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10521
Reviewed-by: Sandaruwan Gunathilake <sandaruwan@readyset.io>
Tested-by: Buildkite CI