replicators: propagate snapshot starting event to controller
During initial snapshot, we don't allow migrations until the snapshot
is completed. The same logic is not applied during resnapshot.
There might be race conditions in between the time we snapshot the
table structure, view structure, migrate new caches, and then start
snapshotting the data. This might lead to the cache with empty rows.

Fix this by propagating the snapshot starting event to the controller.

Fixes: REA-5986

Change-Id: I88131d42a4cfdead9be9cded0aa32abf17d568b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10532
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>