readyset-logictest: Reduce verbosity of error logs on fail
When fuzzing, we can generate 10+ MB test scripts, which are sometimes
included in the error message depending on how it failed. When that
happens, there is nothing else interesting in the error message. Since
the test case is already written out, let's only log the first 256 bytes
of the error message at ERROR level (the full message and script is
still written out at DEBUG level).

Change-Id: I6a6a6964df5ae7b974c3d8597ada07f407c7d6b9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10576
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>