readyset-adapter: Require a Refresh bound on QueryResult
The new trait isn't used yet, so there are no functional changes here.
In the future, we'll need the ability to unpack an upstream result into
DfValue format without simultaneously sending the result to a client.
We move the related MySQL code into the trait implementation so we'll be
able to use it from the adapter.

Change-Id: I2919d09f47663e07f791ca674dc0da624c03f5c6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10605
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI