readyset: Verify estimated snapshot size at startup
When starting Readyset, try to determine how much space a snapshot of
the upstream will consume, and fail the startup verification if it looks
as though we won't have enough space to accommodate the entire snapshot.

Our verification relies on a few assumptions:

- The reported sizes of the upstream tables are basically correct
- All databases (MySQL) or schemas (PostgreSQL) will be replicated
- The tables will take about twice the space when stored in Readyset

Skip this verification check if the user has specified any tables via
the --replication-tables or --replication-tables-ignore arguments.  We
can consider a more granular approach in the future.

Fixes: REA-5937
Release-Note-Core: Add pre-snapshot verification checks to help find
  potential configuration issues earlier. The argument --verify can be
  used to run only the verification checks and then exit. The argument
  --verify-skip can be used to skip the verification checks and continue
  with normal startup.
Change-Id: I1a59c7bad465f87b59f1c1ce6062616d779579bb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10601
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI