readyset-adapter: Fix parameterized LIMITs with TopK flag
Allow Readyset to create caches with parameterized LIMITs when TopK
feature flag is enabled.

This also considers the case when a same query is cached multiple
times with different types of LIMITs (literals and parameterized)
by doing dual lookups, where the adapter tries to find a cache
with the literal form first and falls back to the parameterized
LIMIT if the first lookup fails.

Change-Id: Id4ef724994290240312529530cdf3ddade5637c9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10618
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>