readyset-e2e-tests: Add test for aggregation return types
We don't really check that we match upstream anywhere, and it seemed
unfortunately very cumbersome to add one in core. (E.g. to compare the
column types to upstream in the adapter after preparing a query, though
we should probably do that eventually; we should also probably add
column types to the checks done by the query sampler.) By comparison, it
was much simpler to add a test suite which would have caught this bug
(fixed in the next CL).

Addresses: REA-6022
Change-Id: I93e6a81631d2215a9a2a35b3627eff1c6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10622
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI