readyset-sql-passes: Split adapter rewrites into two phases
In the first phase, we only perform rewrites that leave the AST in a
state such that it is on its own equivalent to the original AST.
Importantly, we perform autoparameterization and normalize the
placeholders.

In the second phase, we factor information out of the query, leaving the
AST changed in a way such that it is NOT equivalent to the original
unless paired with the factored out information (the DfQueryParameters).

Change-Id: I298e03343c55256f49513f65e294d8555266f1b8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10712
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>