benchmarks: Real-time prometheus metrics
Adds an optional mechanism to publish latencies/counts directly as
prometheus metrics, so they can be scraped by prometheus for real-time
display in grafana. This change updates the metrics after each
operation, as oppossed to the existing batch and report in a
background thread.

There was a lot of concern when the benchmarks crate was written wrt
the unintended latency effects of updating atomics that back the
metrics instances. This CL preserves that concern and makes the
immediate reporting of metrics optional, contingent on setting the
`--prometheus-push-gateway` program argument (a pre-existing flag,
which currently reports ... almost nothing to the push-gateway).

The additional value of this work is that it is far easier and more
rational to create grafana dashboards on the continually updated
metrics vs the batch style reporting we have now. The batch style
reporting will create very uneven graphs (as we report only every few
seconds), which makes a poor graph experience.

Change-Id: Idf07d9de3aa9dec12a88880920289f0a435f3e9e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10658
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>