readyset-tracing: Log panics via tracing::error! instead of stderr
This will put panics in the same place as other logs in deployments with
a logfile configured, as well as allow us to inspect panics when using
structured logging (e.g. Antithesis).

We don't enable this in tests (unless running in Antithesis), only
because it breaks the nice colored ASCII output of `pretty_assertions`.

Change-Id: I3779092ff1f8e7f3a1d7fca7eca78c236a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10669
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI