readyset-logictest: Fix no-in-process-readyset build
The newly improved QueryDestination check was using a reexport of the
underlying enum from readyset_client_metrics via readyset_adapter, the
latter of which we don't want to compile in Antithesis builds. Just use
the original enum. (Whether that is a good place for that enum to live
is another discussion.)

Change-Id: Ic6ac7014a4ceaa4888bf1619c2f308786a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10668
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI