readyset-adapter: Preserve query mutations in TopK dual lookup
There was a bug where we passed a clone of the ViewCreateRequest to
the lookup_topk_cache function, and returned its results back if
it succeeds.

The issue is that we returned the results of the rewrite on the
clone, not the original query, so the server received an un-rewritten
stmt, causing an issue with the query matching logic.

Change-Id: I24d7250f6dbdbeb403dad77ce2b868a5054fec63
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10681
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>