dataflow-state: Fix potential deadlock in index creation
There were a few potential deadlocks hanging around within the accesses
through the PersistentStateHandle, and we got bitten by one of them
during the index creation process.

The PersistentStateHandle uses the read-write locks supplied by the
parking_lot crate.  In an attempt to be fair, these read lock requests
will block if the write lock has been requested.  This is a problem if
you are already holding a read lock and then ask for another one.

The deadlock looks something like this:

1. Acquire a read lock
2. Someone else requests the write lock
3. With the read lock held, request another read lock

Since our index creation is parallel and takes multiple read locks in
this manner, it was especially susceptible to this kind of deadlock.

Fortunately, these read-write locks also implement the read_recursive()
interface, which guarantees that subsequent requests for the read lock
will be granted, regardless of waiting writers, with the caveat that
its overuse could cause writer starvation.

There are numerous, and sometimes surprising, places that take a read
lock via the PersistentStateHandle's inner() method, including the
problematic usages during index creation.  Amend the inner() interface
to instead take the read lock via read_recursive(), guaranteeing that
subsequent read lock requests will be granted immediately.

Add a new inner_fair() interface to the PersistentStateHandle, and call
it from the places where blocking should be fine (typically at the
entrance of a critical section).  This interface is simply what inner()
used to be, requesting the read lock via read().

Fixes: REA-6105
Release-Note-Core: Fix a potential deadlock in index creation.
Change-Id: Idc410d7d19f14c7ead40d591a49aa21deae13eef
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10684
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>