adapter: Set upstream version in handshake messages
Upon new client connections, Readyset is setting the default database
version string into the connection handshake packets. When we
introduced the `LazyUpstream`, we made some functions not go upstream
and just return a default value; `version()` is one of those
functions.

The reasoning behind the lazy connects is that when control
plane/operational tools connect to Readyset for health
check/diagnostic information, we were inadvertently opening a lot of
connections to the upstream postgres itself, and thus naively
consuming the limited connections that one can make to PG. Also, if a
given client connection will never need to query the upstream anyways,
why create that connection?

This patch adds a new field to the `UpstreamSystemProperties` struct
for `db_version`, and we pass that to the adapter `Backend`, via it's
builder. `Backend::version()` will return that value if it is `Some`;
else, it'll do it's normal behavior (which may be to return the
default).

Fixes: #1562
Release-Note-Core: Set the correct version of the upstream database
  new connection handshake for both mysql and postgres.
Change-Id: Ib4cd2e85a4271c645e90eb9fd44befdfeff202df
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10703
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>