replicators: remove extra space in snapshot query comment
This commits removes the hardcoded space between the initial comment
marker and the comment itself. This allows for more flexibility in the
comment string as the user can now include special hints like:

```
/*!80030 COMMENT */
/*+ COMMENT */
```

This commit also modifies the get_queries method to not require a
readyset_client::Table as argument. This was used only to extract the
table name and schema inside the function. A relation will be enough
and make unit testing easier.

Added relevant unit tests for the get_queries method.

Fixes: REA-6102

Change-Id: I484c86b6128e9ca2e5d8ce10299a74f331f43dd9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10704
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>