readyset-sql-passes: Include scalar optimization pass in adapter
Adding a few tests to check it's working, just because nothing else was
effected.

We also needed to update how `Expr::Row` was handled in constant
folding, otherwise it would become a string literal representing an
array (e.g. `'{1,2}'`). This would cause autoparameterization to
parameterize out the string literal, which would break `WHERE IN`
handling; i.e. instead of `WHERE IN ((1, 2), (3, 4))`, it would become
`WHERE IN ('{1,2}', '{3,4}')`, and thus `WHERE IN (?, ?)`, when what we
want is `WHERE IN ((?, ?), (?, ?))` (which is later pulled out for
special `WHERE IN` handling by `collapse_where_in`).

Now, we only recurse into the elements of the row expression during
constant folding, leaving expression lowering and evaluation unchanged.

Addresses: REA-5933
AI-Use: level:1
Change-Id: I0603ecd6c7d6759f6aa0837cb19e566a6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10718
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Tested-by: Buildkite CI