ddl_vertical_mysql: Fix handling of DROP VIEW
As the test runs DDL to modify the upstream, it updates a state machine
in an attempt to keep track of the changes it's making via SQL.

Fix an inconsistency in the test's handling of DROP TABLE.  The test was
written to append the CASCADE keyword to the DROP TABLE statement, with
the expectation that doing so would also drop views dependent upon the
table.

MySQL, however, does not implement CASCADE for DROP TABLE (although it
does silently eat the syntax).

Update the SQL the test runs to first drop the dependent views before
dropping the table in order to keep the database and model state
synchronized.

Also add the regression seed to the list of regressions.

Change-Id: I664212f1660af1aca5e9190c67cd172aef1c2232
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10708
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI