Readyset-dataflow: Refactor remapped keys
Remapped keys is stored in a nested hashmap structure. This is complex
to read and understand. This refactor replaces the nested hashmap with
Two structures that compose the remapped keys.

We have a RemappingKey structure to identify a unique remapping.
And a RemappingInfo structure to store the downstream keys and the tag
that triggered the remapping.
Both operations (insert and remove) are kept as O(1) operations for
faster insertion and lookup/removal.

Change-Id: I6ec89cd08cfc9b73c62213012843cad1c169a642
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10743
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI