.buildkite: Allow setting LOG_LEVEL for logictest runs
We also default to info, which is what it was before the recent
logictest logging renovation. It's not too verbose us, just for
Antithesis.

Additionally, for the integration build logictests, provide a single
script to run them all, to make it easier to run them locally.

Note that I don't anticipate wanting to do any of same for the generated
tests run in nightly; and their log level is already configurable if we
really need it, but there are many more so I'm leaving them at error.

Change-Id: Iff5b8d0156f90ed0e4a5087bbd4fb8526a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10751
Tested-by: Buildkite CI
Reviewed-by: Ron Hough <ron@readyset.io>