dataflow-state: fix evict_keys to indicate if anything was evicted
There are situation where we might mark a key as filled but it doesn't
have any rows, like in an empty resultset. In this case, we were
evicting (marking the key as a hole), returning 0 rows and returning 0
bytes evicted. Even though we evicted 0 bytes, we should still
propagate this eviction downstream, otherwise we might end up in a
situation where a downstream node has a filled key with no rows, but
the upstream node has a hole for that key.

Change-Id: I8de0053b6405d14ba53e1a46e7b910c755030877
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10768
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>