dataflow-state: fix evict_keys to indicate correct bytes evicted
A state with a key filled occupies bytes, even if no rows are present.
When evicting from a state, we should take this into account when
reporting bytes freed, otherwise we might end up in a situation where
a downstream node has a filled key with no rows, but the upstream node
has a hole for that key.

Change-Id: I3256a3a25e962dddb432b286d3e71f28d2b020a5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10769
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI