readyset-dataflow: Gate downstream evictions only if we evicted.
When evicting from a node, we unconditionally trigger downstream
evictions. This might cause issues with Straddled Joins. For example,
If we SJ node has other partial materializations above it, and we get
two subsequent evictions, the first in the SJ itself, we will remove
the remappings and evict fine. Now if we receive the second eviction
on a node above the SJ, we will no longer have the remappings to
trigger the eviction downstream. This second eviction is not
necessary if we take into account if we have evicted something.

This works because now evict_keys also returns the amount of bytes
freed taking into account state that has a key filled but no rows.

Fixes: REA-6127
Change-Id: If462b70d45d3adf00186d402a02c8b74651f3527
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10770
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI