readyset-adapter: Return SQL rewrite errors when there's no upstream
Stop swallowing adapter rewrite failures when no upstream is configured
so that errors from SQL passes (e.g. unsupported window functions in
`WHERE` clauses) now surface the original error instead of only erroring
about "no upstream". (We will still get a vague "invalid upstream" error
if we are in a proxying state, such as in a transaction, but don't have
an upstream; I don't know if that's possible, but if it happens it's
probably a bug.) This keeps proxy behaviour intact when fallback exists
while letting tests and standalone deployments see the actual underlying
reason we can't use a cache.

Change-Id: Ie81b5f802b1d2a3baf0bcabe7a5c6aa96a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10772
Tested-by: Buildkite CI
Reviewed-by: Vassili Zarouba <vassili@readyset.io>