readyset-adapter: Adjust type of UpstreamDatabase::ExecMeta
For the purpose of Postgres result format translation, we want the
ability to make a new one of these things that can be used like the
original.  To do that, we introduce a new associated type
ShallowExecMeta: Borrow<ExecMeta>, which is an owned type, but then we
have to adjust the type of ExecMeta so that bound is actually true.
Along the way, we can delete the lifetime parameter from ExecMeta, which
IMO is an enormous improvement just by itself.

Change-Id: I1e0cf3610f612160ae3f1ffc961ef2b73a470868
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10833
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI