antithesis: reduce asserts in builtin lowering
While it was well-intentioned to add antithesis asserts for all of our
builtin functions, it's a bit noisy in the reports as we don't hit
every builtin on every run. This CL reduces the asserts just to make
sure we hit it once for the builtins, not every individual function.

Fixes: REA-6018
Change-Id: I67a0d2a8a1f67dd8ea623ec942c54759de799d58
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10784
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>