readyset-sql-passes: Add a literalize function to generate a query
In shallow caching, we may not have a fully literalized query string
that we can use to refresh a key from upstream, as the query may have
been prepared or submitted via a parameterized extended protocol (i.e.
in Postgres, which is like an unnamed type of prepare).  Given a
parameterized query, this code lets us inject parameters back into the
query so that it can be printed and run.

Change-Id: Ia0a93b26e0fc9fe8a7eb218bcce61a7539a4a401
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10834
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI