readyset-adapter: Check meta compatibility in execute path
In Postgres, the ad hoc path is compatible with shallow cache results
inserted in either ad hoc or execute modes, but execute cannot translate
ad hoc results to execute format.  The adapter will now treat an
incompatible cache entry as a miss and proxy via the extended protocol
to insert a result that works for both ad hoc and execute modes.

Change-Id: I87677e5bde67db42c4ebc38199c837e594bde357
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10839
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>