readyset-adapter: Create shallow cache refresh workers once per adapter
We were creating a pool of shallow cache refresh workers once for every
new client connection, which could end up creating a lot of backend
connections.

Instead, create the pool of workers once in the adapter and pass a copy
of the sending side of the channel to the Backend struct we create for
each client connection.

Also hoist the shallow cache manager out of the connection loop in the
TestBuilder's incarnation of the adapter, sharing it across client
connections.

Change-Id: If65d47bf1dd1e0c136542413de334b59dcb6489f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10828
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>