single_state: Allow to mark a key as filled multiple times
SJ are somehow special. We might be replaying from a miss on the lhs
when we do a lookup on the rhs that also misses. When that happens,
we trigger an upquery on the rhs using a combination of the join key
and the rhs predicate key(s). When the right hand side notifies us
that the replay has completed, we will redo the whole replay,
replaying the lhs and looking up the rhs, this time it should not
miss.

The lhs will be marking the key as filled multiple times, thus we
should allow it.

We might want to optimize this at some point, by resuming the replay
from the segment that missed, instead of replaying again from src.

Change-Id: I3cd1fe2d5e90ac64fad943a48b2c3afbdae351a1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11184
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>