processing: Add MissReplayKey::Explicit
In SJ when we miss, we don't have the all the data needed in the
replayed row to infer the other side key(join + predicate).
Instead we have missing_upqueries that we record the predicate on the
other side.
When building a miss, the other side lookup is derived from the row
we are processing, by extracting the on column from the current side.
However, since the predicate key is not present on this side row, we
cannot build it.

This commit adds a new variant to MissReplayKey that allows us to
specify the key explicitly.

Change-Id: I5339e6e4e879f3f2fdfc0a3e606e4d81dd89eb71
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11185
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>