readyset-logictest: Fix building without in-process readyset
In reviewing the previous change (Ia9499e5c4), I forgot that we don't
even have `readyset_adapter` when building without in-process readyset
for Antithesis. The solution is to push the usage of `MigrationMode`
further down the call stack into the `in_process` module, leaving the
out-of-band determination as just a boolean until then.

Change-Id: I1a1425f10a8df79c1bfad8827084a2fd6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10843
Tested-by: Buildkite CI
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>