readyset-adapter: Fall back to upstream if `rewrite_equivalent` fails
Though we currently don't usually return an error from this set of SQL
rewrite passes, it will become routine in the future as we move some
server passes into the adapter. For example, `expand_stars` will fail
here if we try to rewrite a query before a newly created table is
replicated. Today, in that scenario, we simply won't find a matching
cache and will proxy, but with the new adapter rewrite regime, we will
error out of the rewrite pass and, prior to this fix, return an error to
the client without proxying.

Note that codepath will likely need to be improved later as part of
REA-6109, where we may want to retry rewrite passes. Depending on where
we decide to put the "did we race with DDL?" check, we may return
something other than a `ReadySetResult` from the SQL passes, and that
could be used to determine whether it's a bug or something (should error
like we do today) or likely routine (like can't rewrite without missing
DDL info, should fall back to upstream).

For now, this change allows us to proceed with the initial rewrite pass
moves.

Addresses: REA-5933
Change-Id: Ibf45840fbb7d2a437dc734da5c1e00dd6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10873
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>