replicators: Add support for `RENAME TABLE`
Implements support for `RENAME TABLE` (MySQL) and somewhat improves
`ALTER TABLE ... RENAME TO ...` handling (MySQL and PostgreSQL).

This is achieved by adding a new `AlterTableDefinition::RenameTable`
variant and, when constructing a `Changelist` for the server, converting
MySQL `RenameTableStatement` into a series of `AlterTableStatement`s.

The resulting `Change`s are handled in the replicator by triggering a
resnapshot of each affected table. This is not very efficient, but
actually "renaming" already-existing base tables would break a lot of
assumptions and require coordinating with the worker/domain to rename
the RocksDB directories safely. So this is the safest option, and it
should be good enough for now.

Release-Note-Core: Added support for `RENAME TABLE` and `ALTER TABLE ...
  RENAME TO ...` DDL statements. ReadySet will now detect table renames
  and automatically resnapshot the affected tables to maintain cache
  consistency.
Project: rename
AI-Use: level:3
Change-Id: I275c12bfbf0dfb94fd3a88f55ea3dd896a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10887
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI