replicators: Make drop for resnapshot succeed if table doesn't exist
I initially didn't include this in the rename patch (I23d996cd) because
my tests passed without it and I wasn't sure it would impact other
functionality. Unfortunately, I hadn't automated the specific case we
care about where the _tmp table being dropped was ignored, so I had only
tested that case *with* this change included.

I looked at the code again, and this should be fine for the previous use
case (manually triggered resnapshot) and is necessary for this case.
Also adding an automated test which would have caught this in the first
place.

Change-Id: I8ad310a11ec53fecb0cda17a516e2aac6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10899
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>