readyset-adapter: Consider shallow caches when handling duplicates
Caches are considered duplicates if they have either the same query_id
or the same name.

Previously, the adapter would first drop existing deep caches when a
duplicate deep cache creation was requested.  We now also check for
duplicate shallow caches and drop those as well.

In a similar manner, we now check for duplicates when creating shallow
caches, first dropping them as well.  This makes the drop-duplicates
behavior uniform and unconcerned with whether a cache is deep or
shallow.

Fixes: REA-6154
Change-Id: I01203c75a7dcda034006992ddc93316c40362562
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10905
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI