sql-rewrites: Non-correlated array constructor subqueries
Previously we handled all join conditions for non-correlated array
constructor subqueries as correlated; that is, we always added `ON
TRUE` as the join condition. That was incorrect for non-correlated
subqueries. Further, downstream rewrite passes take care of the join
conditions for LATERALs, so we just need to emit an `Empty` join
condition.

Fixes: REA-6156
Change-Id: I86a1e520862e832315e4d9dc0964e13475ff0792
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10900
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Tested-by: Buildkite CI