mysql-srv: Pre-size reused output buffers
Pass a `size_hint` to `PacketConn::get_buffer()`. This allows better
fit for the requested buffer. Additionally, there was a bug with how
we sorted the buffers: `PacketConn::len()` was looking at the number
of elements currently in the buffer (the `len()` function) and not the
capacity of the buffer. That's fixed in this CL.

This CL also clarifies that we do not re-enqueue
`QueuedPacket::Raw`. It was slightly obscured in
`return_queued_to_pool()`: the `retain()` filtered it out but then the
following iteration kinda ignored it. The same for `get_buffer()`: I
kinda ignored the raw packets. This was confusing, and I'm partly
responsible for it; thus, I'm being far more explicit about by
returning `unreachable!`.

Along the way, this CL batches the column definition packets into one
buffer (`write_column_definitions()`).

Addresses: REA-5695
Change-Id: Id4cae40c529167a2ae69c858f2e901f643c99aad
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10917
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>