dataflow-state: Remove serialized key injectivity test
Numerics and floats sometimes compare as equal because we cast the
numeric to a float and then compare the bits. But there is no reason
their serializations (which include the type information) should be
equal. Lookups are not affected by this because we do
`DfValue::coerce_for_comparison` before comparing and
`DfValue::maybe_coerce_for_table_op` before serialized comparisons. So
this whole test is pretty meaningless.

Fixes: REA-6174
Change-Id: I3ab28116a94075b534696097ecf30f6e6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10929
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>