readyset-dataflow: Coerce window function `SUM` to output type
In cases where the output type differs from the input type (e.g. if `x`
is a `BIGINT`, then `SUM(x)` is `NUMERIC`), normal aggregates handle
this by initializing a zero value with the correct output type and
adding to it. The way window function processing is written, we start
with the input values, in this case a `DfValue::Int`, and never convert
it to the output type. In the bug that exposed this, we would `PREPARE`
the statement with the correct output type (`NUMERIC`), but then when
writing out resultsets, we would attempt to convert a `DfValue::Int` to
a Postgres `NUMERIC` which is invalid and error. We didn't see it, but
this could also overflow during accumulation.

Now, we coerce the sum as necessary. We do so at each step, which could
possibly be avoided (e.g. just convert the first value we see), but I
didn't want to make the code more complex until verifying this issue and
getting it reviewed.

In order to test this, we also start to fill a major testing gap: Rust
e2e tests of window functions. Now that we can plumb a `ParsingPreset`
through `TestBuilder`, we adapt the existing `aggregation_type` test to
support window functions, including verifying identical values are
output. (We could probably add output value testing to the non-window
test cases in this file, but those are better tested.)

Fixes: REA-6125
Change-Id: I60bbfeeaf0f34a7cf6b7fe5b0e76b6486a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10947
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI