replicators: Support `CREATE TABLE ... AS SELECT` on MySQL
While inspecting the MySQL statement replication handling for `CREATE
TABLE ... LIKE` statements, I went looking for other kinds of `CREATE
TABLE` statements that would show up in the binlog *without* the
`UpdatedDbNames` status variable. An AI suggested that `CREATE TABLE ...
AS SELECT` was one such. But when I tested it, it turned out MySQL
actually logs it as the resulting regular `CREATE TABLE` (i.e. with
columns etc., and without the `AS SELECT` clause), followed by row
events for the selected rows to insert them into the new table. However,
it also appends a `START TRANSACTION` option to the `CREATE TABLE`
statement, which we were not parsing. The combination (no
`UpdatedDbNames` and failed to parse) meant we were silently ignoring
the `CREATE TABLE` statement.

First, this CL fixes the parsing for both nom-sql and sqlparser (albeit
with the somewhat hacky solution of producing the AST as a
space-separated key = START, value = TRANSACTION pair, but we are
ignoring it anyway.

Second, we now parse all statements in MySQL replication and attempt to
handle them, regardless of whether `UpdatedDbNames` is set. However, we
still check the status variable to decide whether to error out if
parsing/handling the event fails.

The main caveat is we don't support it on Postgres, and I don't yet know
what that would require.

Addresses: REA-4434, REA-3859
Release-Note-Core: Support replicating MySQL tables created with
  `CREATE TABLE ... AS SELECT`.
Change-Id: I36eaeb8c2823d256da150d7f67b112266a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10969
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>