readyset-adapter: Update more state when dropping shallow caches
When dropping shallow caches, we now update the corresponding entry in
the query status cache.  Refactor some of the logic common between deep
and shallow caches.

Dropping a shallow cache by name now appropriately returns the query to
the SHOW PROXIED QUERIES list in the same manner as deep caches.

Change-Id: Ibc8472e7a6193c99a837e04934cfae99b0572c85
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10973
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI