readyset-adapter: Do not wait on pending during EXPLAIN CREATE CACHE
If a query's migration state was pending, we used to wait around for the
migration handler to eventually perform the dry run migration for us in
order to report whether or not the query is supported.

Instead of waiting around, perform the dry run migration and report the
supported status immediately.

We previously considered any error to be potentially transient if its
caused_by_unsupported() method returned false, which excluded a large
number of errors that do actually indicate our lack of support.

Such errors would leave us in a pending state for 15 minutes (the
default value for --max-processing-minutes) before the migration handler
eventually got around to marking the query as unsupported via timeout,
masking the original error.

Use the new is_transient() method to specifically call out the errors
that are transient, and consider all other errors an indication that we
do not support the query.

EXPLAIN CREATE CACHE should now never report that query support is
pending unless --feature-placeholder-inlining is enabled or we
experience a transient error during the request.

Also update the logic in the background migration handler task to
reflect these same semantics.

Change-Id: I5d962532eb07ce0a35e0275a542082a0a0f76464
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11012
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>