readyset-dataflow: Trigger on_eviction hook inside downstream evict.
Straddled Joins are sensitive to multiple requests for the same
remapped keys. We lookup remapped keys in downstream_dependent_paths.
TopK needs to trigger on_eviction hook on the destination node, but
it was doing it in handle_eviction_bytes/keys and also calling
downstream_dependent_paths. During trigger_downstream_evictions, we
also need to resolve all downstream paths, which was causing
remapped keys to be missing.

This commit adjusts the logic to call on_eviction hook inside
trigger_downstream_evictions.

Cherry-Picked test case from Mohamed:

50f799041cecfb83fb45d41f868ed78fc2382802

Add a test that makes sure that the on_eviction hook is called
correctly in TopK node.

Switch `flush_partial` endpoint to POST so we can use it in tests,
the endpoint is only called using cURL, that's why it was
originally set to GET.

Change-Id: If8b7dad07ab7f95517c53476c5017c36c1832044
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10979
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI