readyset-dataflow: Fix unquery eviction for straddled joins.
During on_replay_misses, we add remapped keys for left and right
ancestors if the node is partial. However during unquery eviction,
we were triggering evictions for all ancestors that were not a
base node. This condition is not guaranteed to be the same, as a node
without a state for example, is not a base node but is also not
partial, in this situation we would trigger eviction for a node that
does not have a remapping, causing the downstream nodes to not find
a tag to fullfill the eviction key.

Fixes: REA-6145

Release-Note-Core: Fixed an issue during eviction of straddled joins
  causing it to attempt to evict a key that does not have a remapping.
Change-Id: I86ea30e7cc5881f2fa895f4d872ec0264211bca4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10985
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>