readyset-adapter: Remove questionable error discrimination
If we fail to read from a view, we put the view into a collection of
failed views that helps us maintain a cache of views that are still
valid.

Back in I0a17c4e6, we tried to differentiate some errors from others
when deciding if a view should go into the collection, all in the name
of performance.  The logic here was strange, considering most errors as
insufficient cause to fail a view, while considering others such as
networking errors as sufficient to fail a view.

This doesn't make much sense, so consider any error sufficient to remove
it from our cache of views.

Change-Id: I8e25927db4974ba6494174a1462d91e9cdfc2b8f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11007
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>