readyset-logictest: Allow comparing results as JSON values
This allows us to ignore formatting, key ordering, etc.

This may not be universally desirable (i.e. it would generally be
preferable to compare JSON values as strings and match upstream's
storage and formatting), but this makes development of JSON related
features much easier. Furthermore, you might be surprised to find out,
we do in fact not match either MySQL or Postgres in our output
everywhere.

This allows us to run the `json_object` test against upstreams, not only
Readyset, which of course reveals that (1) it only works against MySQL;
and (2) our implementation of `json_build_object` does not match
Postgres. So the test is moved into the mysql directory and a subsequent
patch fixes the Postgres behavior.

AI-Use: level:2
Change-Id: I8a62a7e9ead6ec8bc8c3fd06276958e56a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11001
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI